Removes defmt from radio tests and moves ble test under test/ble for consistency
